+++ attributemap/microsoft2name.php
@@ -1,0 +1,10 @@
+<?php
+
+$attributemap = array(
+    'microsoft.displayName' => 'displayName',
+    'microsoft.id' => 'uid',
+    'microsoft.mail' => 'mail',
+    'microsoft.surname' => 'sn',
+    'microsoft.givenName' => 'givenName',
+    'microsoft.name' => 'cn',
+);