+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+_Release: 2018-?
+
+* Template for YahooOIDC, MicrosoftOIDC, LinkedIn and Facebook
+* Add support for enabling http request/response logging
+* Add general debug information
+
+## v1.0.0
+
 _Released: 2018-?_
 
 * Generic OAuth2/OIDC module